Code tefactoring

+++ src/components/cards.js
@@ -3,6 +3,9 @@ export {cardsList, initialCards, addCard, createCard}
 
 /*Импорты*/
 import {openPopup} from "./popups.js";
+import {deletePlaceCardApi,
+        likePlaceCardApi,
+        dislikePlaceCardApi} from "./api";
 
 /*Массив с карточками*/
 const initialCards = [
@@ -42,8 +45,20 @@ const cardTemplate = document.querySelector('.template').content;
 const popupImgWatch = document.querySelector('.popup_img-watch');
 
 /*Декларируем функцию работы с лайком*/
-function likeCard(likeButton){
-    likeButton.classList.toggle('cards__card-like-button_active');
+function likeCard(likeButton, card, cardLikeBtnLabel){
+    if (Array.from(likeButton.classList).includes('cards__card-like-button_active')){
+        dislikePlaceCardApi(card._id)
+            .then(res => {
+                likeButton.classList.remove('cards__card-like-button_active');
+                cardLikeBtnLabel.textContent = res.likes.length;
+            });
+    }else{
+        likePlaceCardApi(card._id)
+            .then(res => {
+                likeButton.classList.add('cards__card-like-button_active')
+                cardLikeBtnLabel.textContent = res.likes.length;
+            });
+    }
 }
 
 /*Декларируем функцию удаления карточки*/
@@ -52,21 +67,48 @@ function deleteCard(deleteButton){
 }
 
 /*Декларируем функцию модификации клона карточки*/
-function createCard(cardName, cardLink){
+function createCard(card, user){
     const newCard = cardTemplate.querySelector('.cards__list-item').cloneNode(true);
     const newCardImage = newCard.querySelector('.cards__card-image');
     const popupImage = document.querySelector('.form-img-watch__img');
     const popupImageTitle = document.querySelector('.form-img-watch__img-title');
+    const cardDeleteBtn = newCard.querySelector('.cards__card-delete-button');
+    const cardLikeBtn = newCard.querySelector('.cards__card-like-button');
+    const cardLikeBtnLabel = newCard.querySelector('.cards__card-like-button-label');
 
-    newCardImage.setAttribute('src', `${cardLink}`);
-    newCardImage.setAttribute('alt', `${cardName}`);
-    newCard.querySelector('.cards__card-name').textContent = cardName;
-    newCard.querySelector('.cards__card-like-button').addEventListener('click', (evt) => {
-        likeCard(evt.target);
-    });
-    newCard.querySelector('.cards__card-delete-button').addEventListener('click', (evt) => {
-        deleteCard(evt.target);
-    });
+    /*настраиваем картинку в карточке*/
+    newCardImage.setAttribute('src', `${card.link}`);
+    newCardImage.setAttribute('alt', `${card.name}`);
+    newCard.querySelector('.cards__card-name').textContent = card.name;
+
+    /*Настраиваем кнопку лайк*/
+    if (card.likes.length !==0){
+        card.likes.forEach(cardLike =>{
+            if (cardLike._id === user){
+                cardLikeBtn.classList.add('cards__card-like-button_active')
+            }
+        })
+    }
+
+    cardLikeBtnLabel.textContent = card.likes.length;
+
+    cardLikeBtn.addEventListener('click', evt => {
+        likeCard(evt.target, card, cardLikeBtnLabel);
+    })
+
+    /*Проверяем совпадение пользователей и настраиваем кнорку "Удалить"*/
+    if (card.owner._id === user){
+        cardDeleteBtn.addEventListener('click', (evt) => {
+            deletePlaceCardApi(card._id)
+                .then(() =>{
+                    deleteCard(evt.target)
+                });
+        });
+    }else{
+        cardDeleteBtn.classList.add('cards__card-delete-button_inactive')
+    }
+
+    /*Добавляем возможность открытия карточки по клику на картинку*/
     newCardImage.addEventListener('click', (evt) => {
         /*Синхронизируем данные карточки и открываемой формы*/
         popupImage.setAttribute('src', evt.target.src);
